requirements.txt
openai>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
gitpython>=3.1.40
click>=8.1.7
colorama>=0.4.6
numpy>=1.24.0
scikit-learn>=1.3.0
rich>=13.0.0
prompt-toolkit>=3.0.0
tabulate>=0.9.0
tqdm>=4.65.0